Update _index.html, fix cordova routing bug
Regarding this issue https://github.com/angular/angular/issues/22509
There is a bug on mobile with cordova when navigating back to the previous page. The router-outlet is appending rather than replacing the component, it is very annoying. 
It seems that there isn't any solution now other than reseting the addEventListener and removeEventListener functions.